requirements.txt
numpy==1.21.5
pandas==1.4.2
PyYAML==6.0
scikit_learn==1.1.3
torch==1.11.0
transformers==4.19.2
black==22.3.0